feat: added expense-category-in-use deletion block

+++ src/main/java/com/jadebloom/goblin_api/expense/service/impl/ExpenseCategoryServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import com.jadebloom.goblin_api.currency.error.CurrencyInUseException;
 import com.jadebloom.goblin_api.expense.dto.CreateExpenseCategoryDto;
 import com.jadebloom.goblin_api.expense.dto.ExpenseCategoryDto;
 import com.jadebloom.goblin_api.expense.dto.UpdateExpenseCategoryDto;
@@ -14,6 +15,7 @@ import com.jadebloom.goblin_api.expense.error.ExpenseCategoryNameUnavailableExce
 import com.jadebloom.goblin_api.expense.error.ExpenseCategoryNotFoundException;
 import com.jadebloom.goblin_api.expense.mapper.ExpenseCategoryMapper;
 import com.jadebloom.goblin_api.expense.repository.ExpenseCategoryRepository;
+import com.jadebloom.goblin_api.expense.repository.ExpenseRepository;
 import com.jadebloom.goblin_api.expense.service.ExpenseCategoryService;
 
 @Service
@@ -21,12 +23,17 @@ public class ExpenseCategoryServiceImpl implements ExpenseCategoryService {
 
     private final ExpenseCategoryRepository expenseCategoryRepository;
 
+    private final ExpenseRepository expenseRepository;
+
     private final ExpenseCategoryMapper mapper;
 
     public ExpenseCategoryServiceImpl(
             ExpenseCategoryRepository expenseCategoryRepository,
+            ExpenseRepository expenseRepository,
             ExpenseCategoryMapper mapper) {
         this.expenseCategoryRepository = expenseCategoryRepository;
+
+        this.expenseRepository = expenseRepository;
 
         this.mapper = mapper;
     }
@@ -104,6 +111,13 @@ public class ExpenseCategoryServiceImpl implements ExpenseCategoryService {
 
     @Override
     public void deleteById(Long expenseCategoryId) {
+        if (expenseRepository.existsByExpenseCategory_Id(expenseCategoryId)) {
+            String f = "Cannot delete the expense category with the ID '%d': some amount of expenses depend use it";
+            String errorMessage = String.format(f, expenseCategoryId);
+
+            throw new CurrencyInUseException(errorMessage);
+        }
+
         expenseCategoryRepository.deleteById(expenseCategoryId);
     }
